changed singletons and static variables to follow better encapsulation practices to prevent future unintended side effects.
